// submodules/common-dto/api-client/main/models/UserVerificationStatus.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * User verification status
 */
export enum UserVerificationStatus {
    VERIFIED = 'verified',
    NOT_VERIFIED = 'notVerified',
    IN_PROGRESS = 'inProgress',
    VERIFICATION_FAILED = 'verificationFailed',
}